Adding initial layout for project specs.
As git does not support commiting empty directories and
new specifications will be submitted many levels away from the
root I chose to copy the tempate.rst into each projects directory.

This is something we could remove in future if desired but
seemed like a sensible way to setup the structure for now.

Change-Id: I154fe85f7de1cfffeaf2b2eaea31e821129342a5